requirements.txt
flask
openai
requests
PyJWT
python-dotenv
gunicorn